Issue #7: Remove obsolete Status object.
