+++ src/Messages/Responses/Status.php
@@ -2,6 +2,8 @@
 
 namespace EC\Poetry\Messages\Responses;
 
+use EC\Poetry\Messages\Components\Traits\WithStatusTrait;
+
 /**
  * Status response, sent back to poetry server after a notification has come in.
  *
@@ -9,11 +11,13 @@ namespace EC\Poetry\Messages\Responses;
  */
 class Status extends AbstractResponse
 {
+    use WithStatusTrait;
+
     /**
      * {@inheritdoc}
      */
     public function getTemplate()
     {
-        return 'messages::request';
+        return 'messages::status';
     }
 }
